libs/py/db/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "py_db" ]

[tool.coverage.report]
exclude_lines = [ "if TYPE_CHECKING:" ]
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-report html:'../../../coverage/libs/py/db/html' --cov-report xml:'../../../coverage/libs/py/db/coverage.xml' --html='../../../reports/libs/py/db/unittests/html/index.html' --junitxml='../../../reports/libs/py/db/unittests/junit.xml'"

[tool.poetry]
name = "py-db"
version = "1.0.0"
description = "Automatically generated by Nx."
authors = [ ]
license = "Proprietary"
readme = "README.md"
group = { }

  [[tool.poetry.packages]]
  include = "py_db"

  [tool.poetry.dependencies]
  python = "~=3.11.5"
  python-dotenv = "^1.0.0"
  sqlalchemy = "^2.0.20"
  alembic = "^1.12.0"
  asyncpg = "^0.28.0"
  passlib = "^1.7.4"
  pydantic = "^2.3.0"

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"
